Rename sharedItems service and store
