+++ src/store/storeConfig.js
@@ -39,7 +39,7 @@ import uiModeStore from './uiModeStore'
 import windowVisibilityStore from './windowVisibilityStore'
 import messageActionsStore from './messageActionsStore'
 import reactionsStore from './reactionsStore'
-import conversationSharedItemStore from './conversationSharedItemsStore'
+import sharedItemStore from './sharedItemsStore'
 
 export default {
 	modules: {
@@ -62,7 +62,7 @@ export default {
 		windowVisibilityStore,
 		messageActionsStore,
 		reactionsStore,
-		conversationSharedItemStore,
+		sharedItemStore,
 	},
 
 	mutations: {},
